+++ src/main/java/com/soft2242/shop/entity/GoodsDetail.java
@@ -45,7 +45,7 @@ public class GoodsDetail {
     @ApiModelProperty("逻辑删除(0-未删除，1-已删除)")
     @TableField("delete_flag")
     @TableLogic
-    private Byte deleteFlag;
+    private Integer deleteFlag;
 
     @ApiModelProperty("创建时间")
     @TableField(value = "create_time", fill = FieldFill.INSERT)
