+++ src/main/java/com/soft2242/shop/entity/IndexCarousel.java
@@ -40,12 +40,13 @@ public class IndexCarousel {
 
     @ApiModelProperty("类型（1-首页，2-商品分类）")
     @TableField("type")
-    private Byte type;
+    private Integer type;
 
     @ApiModelProperty("逻辑删除(0-未删除，1-已删除)")
     @TableField("delete_flag")
     @TableLogic
-    private Byte deleteFlag;
+
+    private Integer deleteFlag;
 
     @ApiModelProperty("创建时间")
     @TableField(value = "create_time", fill = FieldFill.INSERT)
